Miscellaneous build improvements
We now force (re)download/install cside-codemirror and cside-choicescript as part of a build. This is annoying but necessary as npm can't seem to tell when there's a new version of a git repo. This meant it was quite easy to accidentally build (or ship) with stale versions of these components.

The concat task will now cause a build failure rather than just warning if it can't find a file.

The clean task will now purge everything properly (the glob pattern was missing a few files before).

Updated grunt+nw-builder dependencies.