Add Project Search/Replace (Tab) Functionality
Adds support for searching for, and replacing,
values across a project. This is presented via
a new "Search and Replace" tab.

The functionality is built on top of the Monaco
editor's findMatches API, and has been tailored
to match the in-file search/replace provided by
the same.

That means it supports:

- Case Matching
- Regex Searches
- Replace Patterns

Like the in-file search widget, the UI here
uses textareas in order to allow for multiline
matches (also following its ctrl+enter convention).

Results are displayed by scene, and 'in context'
of the surrounding text. Replace values are
likewise pre-calculated, and displayed in context.

By default the search is 'OnType', but this can
be disabled via a toggle alongisde case matching etc.

For performance reasons Monaco limits the number
of results returned for any model to 999. CSIDE
will present a notification next to the results
to inform the user in such circumstances.

Results will be kept in-sync with the model on edits.
Search results are cached at the Scene level,
so only the Scene(s) being edited will be re-indexed.

We also specify a numerical expand threshold, which
determines when a scene's results are displayed in
the UI (or collapsed by default). Collapsing results
allows us to delay unnecessary rendering and calculation
of the in-context text.