[bug fix] Keep URIs in sync when renaming scenes
The internal Monaco services rely on Model URIs
to uniquely identify and keep track of models.
Unfortunately, there is no way to edit the Uri
of a model once it is created, which is problematic
when renaming scenes.

So what we're doing here is now replacing the model
on a rename. This works well enough, apart from one
really BIG caveat: it causes us to lose the
undo/redo stack. There might be ways to work around
this, but let's consider it a necessary regression
for now (see #115).

An alternative option would have been to use UIDs
for the URIs, over the file-path style, but this
would make it difficult on a lot of the
languageservice features such as the indexer.