[bug fix] Fix Monaco model creation on Windows
284737c moved model creation to be linked
with changes to the Scene path() observable.
This caused an issue on Windows because
the backslashed paths are normalized. This
is enough of a difference for the model creation
to be triggered twice, but not enough of a change
for Monaco to consider them separate URIs,
leading to an error.

This patch leaves the extended path() variable
to handle future changes, but moves initial model
creation back out. This also allowed me to fix
another accidental regression caused by 284737c
which was the loss of the default startup.txt
code.