Move to per-scene indentation
Moves away from a 'global' indentation
setting and allows each scene to have
its own. The type (spaces, tabs) and
size are displayed in the code-footer.
Clicking this will present popups for
changing the values for the currently
selected scene.

On creation of new scenes, the defaults
set in settings will be used. When
loading a scene, we rely on Monaco's
'detectIndentation'.

To further accomodate that, there are now 3
global settings: usespaces, tabsize
and indentspaces. The reason for
separating indentspaces and tabsize
is to support scenarios where your preferred
default space insert value might not
be the same as your desired tab size.

Meaning if we open a file and detect tabs,
it'll apply the default tabsize. Whereas
detecting space indentation will infer
the level from the contents.