Support for multiple editors
Another huge patch (too big, really),
but there was a lot of miscellaneous re-plumbing
required to get CSIDE where it needed to be to
support this.

General description:
Multiple editors are only supported per-project,
that means you can't view two editors from two
different projects simultaneously. This was
necessary as CSIDE has been architected to
always want an "active" or "selected" project.
It would be too difficult to determine which
project to test, run or do anything else with, if
editors could be mixed and matched.

However, you can switch between projects without
losing your 'editor state' (which ones you have
open etc.).

Closing/opening editors has no bearing on whether
a scene is technically open in CSIDE. The single
source of truth for that is still the scene list.
Whether an editor is open or closed for a scene
does determine if it is "selected" and to that
end you can now have multiple scenes selected
across projects simultaneously.

In theory the code supports multiple editors per scene, and
monaco handles this well on the model side. However, for the
time being I figured that this would more confusing than
helpful to most users.

In order to ease users into this functionality, by
default, CSIDE will behave as it has always done:
clicking another scene will seem to swap the editor
to it (whether in the same project or not). Multiple
editors only kick in when:
- The user holds shift and clicks a new scene
- CSIDE wants to automatically open something
  without losing your place (e.g. showing you an issue,
  or search result)

Once you have 2+ editors open, subsequent clicks selections
will also create new editors. This seemed the more sensible
option than trying to guess which editor the use wouldn't
mind being repurposed.

Architecture design:
A CSIDEEditor contains an instance of a monaco editor
and points at a specifc CSIDEScene (and its underlying
monaco model). Editors are managed at the project level,
in order to simplify display and manipulation. An editor
can subscribe and unsubscribe to a project, and will
display based on whether the project is "active". Likewise,
if a scene is "selected" we can follow the dependency chain
and make the editor "active", switching focus to the project
it's subscribed too.

The chain is quite neat, but not without complexity.
For example, there is an 'allEditors' computed variable
that exists solely for cascade applying settings to all
editors regardless of project. You can grab all an scene's
editors via a similarly computed variable (filtering on the
project's editor subscription list).

When a project isn't active the monaco instances are hidden
rather than destroyed or recycled. This isn't great for memory,
particularly if users don't realise they're leaving editors open,
but otherwise more performant than recreating them on every project swap.

The editor "tags" or labels are the only things tied to knockout's MVC
directly (to avoid re-renders tanking the editors), and the editor
dom elements are manually kept in sync with them on re-ordering/closing etc.

Other notes of interest:
- The word "selected" (selectedScene/Project) has
  largely been removed in favour of "active".
- Word count had a major rejig as part of this patch,
  seeing as it now needs to be editor (not scene) centric

Possible future enhancements:
- Manually resizeable editors
- Persistent editor state (i.e. between app sessions)
- Configuration options? Some might want to disable multi-editors?