Support all extensions as CSIDEFiles
Continuing on from:
bd3004a10cc70ccce979c8294fdd2dea3f22acdf

This patch removes hard-coded references
to ".txt" in the management of CSIDEFiles
(previously CSIDEScenes).

A notable implementation change is that
a File's name is no longer manually set
but is derived from the path AND now
includes the file's extension.

When editing the name of a File,
nameInterface abstracts away the extension
and adds it back (making it impossible
for the user to change a File's extension
accidentally).