Improve multi-editors and add pin support
Improves a chunk of the multi-editor
handling, including resizing and
reordering. It also changes editor
allocation to be based on 'pinning'.
By default if the SHIFT key isn't
held down then the focused editor,
or first unpinned editor, will be swapped out.
If the SHIFT key is held, or there are no
unpinned editors, then a new editor will be created.