[bug fix] CMD+W should close editors not the app on macOS
It's long been an issue that a stray CMD+W on macOS can
cause the application to quit out (luckily still triggering
dirty checks etc). CMD+W also no longer seems to be working
against Mousetrap and Monaco editor shortcuts. It's proven
difficult to catch and fix this properly, but this provides
a janky hack to detect if a closure event was triggered whilst
those buttons were pressed, in that case it'll trigger an editor
closure instead (inline with other platform behaviours).

Fixes #1.